Fixed db context factory for migration where access token needs to be applied as password
